feat(11107): Lookup tables page ready
