+++ graylog2-web-interface/src/components/lookup-tables/index.ts
@@ -17,8 +17,8 @@
 import './adapters';
 import './caches';
 
-// export { default as LookupTablesOverview } from './LookupTablesOverview';
-export { default as LookupTablesOverview } from './lookup-table';
+export { default as LUTLayout } from './lut-layout';
+export { default as LookupTablesOverview } from './lookup-table-list';
 export { default as LUTTableEntry } from './LUTTableEntry';
 export { default as LookupTableView } from './LookupTableView';
 export { default as LookupTableForm } from './LookupTableForm';
@@ -26,7 +26,6 @@ export { default as LookupTableFields } from './LookupTableFields';
 
 export { default as LookupTableCreate } from './LookupTableCreate';
 
-// export { default as CachesOverview } from './CachesOverview';
 export { default as CachesOverview } from './cache-list';
 export { default as CacheTableEntry } from './CacheTableEntry';
 export { default as Cache } from './Cache';
@@ -35,7 +34,6 @@ export { default as CacheCreate } from './CacheCreate';
 export { default as CachePicker } from './CachePicker';
 export { default as CachesContainer } from './CachesContainer';
 
-// export { default as DataAdaptersOverview } from './DataAdaptersOverview';
 export { default as DataAdaptersOverview } from './adapter-list';
 
 export { default as DataAdapterTableEntry } from './DataAdapterTableEntry';
