+++ graylog2-web-interface/src/components/lookup-tables/lookup-table-list/unit-test/lookup-table.test.tsx
