+++ graylog2-web-interface/src/components/lookup-tables/lookup-table-list/utils.ts
